% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{A sampling function to make barplots of successes}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the iterations}

\item{n}{the sample size}

\item{p}{the probability}
}
\value{
barplot(s) of successes in each sample
}
\description{
A sampling function to make barplots of successes
}
\examples{
\dontrun{mybin(1000, 20, 0.3)}
}
