% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{A bootstrapping function to take random samples and make histograms}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations}

\item{x}{the sample object of some distribution}

\item{fun}{the function to perform on the same (mean, sd, etc)}

\item{alpha}{variable for finding a certain confidence interval}

\item{cx}{an expansion factor}

\item{...}{}
}
\value{
a histogram of the sample and a list of simulation statistics
}
\description{
A bootstrapping function to take random samples and make histograms
}
\examples{
\dontrun{myboot2(10000,x=sam,fun=function(x) mean(x),alpha=0.05,xlab="mean(x)",col="Red",cx=1.5)}
}
