% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function for Mu}
\usage{
myci(x)
}
\arguments{
\item{x}{the single sample input}
}
\value{
prints the c.i. of x
}
\description{
Confidence Interval Function for Mu
}
\examples{
\dontrun{myci(x=rnorm(35,mean=7,sd=10))}
}
