% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A log likelihood, must be defined.}

\item{x}{Vector of the sample daata}

\item{param}{A regular sequence to and from defined values}

\item{...}{Extra arguments to make output more specific to needs}
}
\value{
A list of max's, params, column sums, and slope and produces a plot
}
\description{
Maximum Likelihood Function
}
\examples{
\dontrun{mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
