% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Function to create plot of the p-value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{The t statistic (Tcalc)}

\item{xmax}{Maximum x-axis value (negative and positive)}

\item{n}{The sample size}

\item{alpha}{the likelihood that the true population parameter lies
outside the confidence interval}
}
\value{
a plot and list of the quantile and p-value
}
\description{
Function to create plot of the p-value
}
\examples{
\dontrun{mypvalue(tcalc, n = 25 , alpha = 0.10)}
}
