% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{A function to read in a csv file}
\usage{
myread(csv)
}
\arguments{
\item{csv}{A file to read data from}
}
\value{
A data frame from the file
}
\description{
A function to read in a csv file
}
\examples{
\dontrun{ddt <- DDT.csv; myread(csv = ddt)}
}
