% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{A function for making a scatter histogram graph}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}
}
\value{
A histogram made from barplots
}
\description{
A function for making a scatter histogram graph
}
\examples{
\dontrun{scatterhist(150, 300, "X Label", "Y Label")}
}
